Add more items

+++ Assets/Scripts/Game/Upgrades/Upgrader.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 public abstract class Upgrader : MonoBehaviour
 {
     [SerializeField] private Wallet _wallet;
-    [SerializeField] private int _price;
+    [SerializeField] private long _price;
     [SerializeField] private float _priceIncrease;
     private int _level;
 
-    public event Action<int, int> Changed;
+    public event Action<int, long> Changed;
 
     public Wallet Wallet => _wallet;
     public int Level => _level;
-    public int Price => _price;
+    public long Price => _price;
 
     public void Execute()
     {
@@ -21,7 +21,7 @@ public abstract class Upgrader : MonoBehaviour
 
         _wallet.ReduceMoney(_price);
         UpgradeTarget();
-        _price = Convert.ToInt32(_price * _priceIncrease);
+        _price = Convert.ToInt64(_price * _priceIncrease);
         _level++;
         Changed?.Invoke(_level, _price);
     }
